requirements2.txt
torch==2.0.1
torchvision==0.15.2
openai-clip-anytorch==0.1.2
grad-cam==1.4.6
timm==0.9.2
Pillow==9.5.0
numpy==1.24.3
matplotlib==3.7.1
opencv-python==4.8.0.74
scikit-image==0.21.0
jupyter==1.0.0
ipykernel==6.23.1